requirements.txt
Flask==1.1.2
gunicorn==20.0.4
imutils==0.5.3
matplotlib==3.3.1
opencv-contrib-python==4.4.0.42
opencv-python==4.4.0.42
pandas==1.1.0
Pillow==7.2.0
python-dateutil==2.8.1
pyttsx3==2.90
Werkzeug==0.16.1
zipp==3.1.0
